requirements.txt
appnope==0.1.0
awsebcli==3.13.0
backcall==0.1.0
blessed==1.14.2
boto3==1.7.28
botocore==1.10.28
cement==2.8.2
colorama==0.3.7
decorator==4.3.0
Django==2.0.5
django-cors-headers==2.2.0
django-extensions==2.0.7
django-storages==1.6.6
djangorestframework==3.8.2
docker==2.6.1
docker-pycreds==0.2.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
parso==0.2.0
pathspec==0.5.5
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.13
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
websocket-client==0.47.0
